flaskapp/requirements.txt
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
dill==0.2.7.1
enum34==1.1.6
Flask==0.12.2
Flask-Cors==3.0.3
future==0.16.0
futures==3.1.1
google-auth==1.1.1
google-cloud-core==0.27.1
google-cloud-language==0.29.0
google-gax==0.15.15
googleapis-common-protos==1.5.2
grpcio==1.4.0
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
oauthlib==2.0.4
ply==3.8
protobuf==3.4.0
pyasn1==0.3.6
pyasn1-modules==0.1.4
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
tweepy==3.5.0
urllib3==1.22
Werkzeug==0.12.2
